requirements.txt
Django==2.2
djangorestframework==3.7.7
gunicorn==19.7.1
pytz==2018.3
six==1.15.0
django-filter==2.3.0
jsonschema==3.2.0

